% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chk0to1}
\alias{chk0to1}
\title{Checks whether bounded between 0 and 1.}
\usage{
chk0to1(X)
}
\arguments{
\item{X}{numeric}
}
\value{
numeric
}
\description{
Many quantities exist only in unit space, that is, within the
range \eqn{[0,1]}.
}
\details{
This function halts operation if range of \code{X} is outside \eqn{[0,1]}.
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}

