% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Elastic.Properties}
\alias{Elastic.Properties}
\alias{bulk_modulus}
\alias{elasticProperties}
\alias{elastic_properties}
\alias{shear_modulus}
\title{Elastic properties from seismic velocities}
\usage{
Elastic.Properties(Vp, Vs, dens)

shear_modulus(Vs, dens)

bulk_modulus(Vp, Vs, dens, mu = NULL)
}
\arguments{
\item{Vp}{numeric; the P-wave velocity of the rock in km/s}

\item{Vs}{numeric; the S-wave velocity of the rock in km/s}

\item{dens}{numeric; the rock density, in kg/m^3}

\item{mu}{numeric; the shear modulus of the rock, in Pa}
}
\description{
Elastic properties from seismic velocities
}

