\docType{package}
\name{hydrogeo-package}
\alias{hydrogeo}
\alias{hydrogeo-package}
\title{Tools for hydrogeology and poroelasticity.}
\description{
  This package provides tools to calculate quantities
  commonly found in hydrogeology and poroelasticity studies
  including, to name a few, hydraultic conductivity,
  permeability, transmissivity, etc.
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\seealso{
  \code{\link{some_function}},
  \code{\link{some_other_func}}
}

