% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogeo.p-package.R
\docType{data}
\name{hydrogeo.p-constants}
\alias{.hg_constants}
\alias{constants}
\alias{constants._hg_}
\alias{get_constants}
\alias{hydrogeo.p-constants}
\title{Constants used as defaults}
\format{An object of class \code{list} of length 6.}
\usage{
.hg_constants

constants._hg_(do.str = TRUE)

get_constants()
}
\arguments{
\item{do.str}{logical; should the structure be printed?}
}
\description{
The hydrogeologic response of an aquifer system depends on its mechanical
and hydraulic properties; if these are not known or 
specified, these constants are used.
}
\details{
The helper function \code{\link{constants._hg_}}
shows (the structure of, optionally)
and returns \code{.hg_constants}. 
\code{\link{get_constants}} simply accesses \code{\link{constants._hg_}}

The following constants are set here:
gravity, properties of water, Poisson's ratios for a typical
drained and undrained material, the compressibities
of fluid and undrained fluid-saturated rock,
atmospheric properties at ATP, and some conversion
factors.
}
\note{
This functionality may be replaced in the future with
the \code{\link[settings]{options_manager}} function on CRAN
(in the  new \pkg{settings} package).
}
\seealso{
\code{\link{hydrogeo.p-units}}, \code{\link{hydrogeo.p}}
}
\keyword{datasets}

