% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{hydrogeo.p-package}
\alias{hydrogeo.p}
\alias{hydrogeo.p-package}
\title{Tools for hydrogeology and poroelasticity.}
\description{
This package provides tools to calculate quantities
commonly found in hydrogeology and poroelasticity studies including,
to name a few, hydraultic conductivity, permeability, transmissivity, etc.
}
\section{Scientific background}{


TODO

\subsection{Physical parameters}{
TODO
}
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
\references{
Brocher, T. M. (2005).
Empirical relations between elastic wavespeeds and density in the Earth's crust.
\emph{Bulletin of the Seismological Society of America}, \strong{95} (6), 2081-2092.

Davies, S. N., and R. J. M. DeWiest (1966),
\strong{Hydrogeology}, \emph{J Wiley and Sons}, \emph{New York}, 61.

Lachenbruch, A. H. (1980),
Frictional heating, fluid pressure, and the resistance to fault motion,
\emph{J. Geophys. Res.}, \strong{85} (B11), 6097-6112,
doi: 10.1029/JB085iB11p06097

Rice, J. R., and M. P. Cleary (1976),
Some basic stress diffusion solutions for fluid-saturated elastic porous
media with compressible constituents,
\emph{Rev. Geophys.}, \strong{14} (2), 227-241,
doi:10.1029/RG014i002p00227

Roeloffs, E. (1996),
Poroelastic Techniques in the Study of Earthquake-Related Hydrologic Phenomena,
\emph{Advances in Geophysics}, \strong{37}, 135-195,
doi: 10.1016/S0065-2687(08)60270-8

Rojstaczer, S., and D.C. Agnew (1989),
The Influence of Formation Material Properties on the Response of
Water Levels in Wells to Earth Tides and Atmospheric Loading,
\emph{J. Geophys. Res.}, \strong{94} (B9), 12403-12411.

Shapiro, S. A., Huenges, E. and Borm, G. (1997),
Estimating the crust permeability from fluid-injection-induced
seismic emission at the KTB site,
\emph{Geophysical Journal International}, \strong{131} (2),
doi: 10.1111/j.1365-246X.1997.tb01215.x

Shepard, F. P. (1954),
Nomenclature based on sand-silt-clay ratios,
\emph{Journal of Sedimentary Research}, \strong{24} (3), 151-158.
}
\seealso{
\code{\link{compressibility}},
\code{\link{dimensional_units}},
\code{\link{hydrogeo.p-constants}},
\code{\link{hydrogeo.p-units}},
\code{\link{hydraulic_diffusivity}},
\code{\link{hydraulic_conductivity}},
\code{\link{permeability}},
\code{\link{porosity}},
\code{\link{skempton}},
\code{\link{storativity}},
\code{\link{transmissivity}}
}

