% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_units.R
\name{hydrogeo.p-units}
\alias{atm2bar}
\alias{bar2pa}
\alias{bbl2cm}
\alias{ft2m}
\alias{hpa2pa}
\alias{hydrogeo.p-units}
\alias{kpa2pa}
\alias{m2ft}
\alias{pa2bar}
\alias{pa2hpa}
\alias{pa2kpa}
\alias{to_cmsquared}
\alias{to_darcies}
\alias{to_millidarcies}
\alias{to_msquared}
\title{Conversion of units.}
\usage{
to_msquared(Perm_Darcy)

to_darcies(Perm_sqm, milli = FALSE)

to_millidarcies(Perm_sqm)

to_cmsquared(Perm_sqm)

bbl2cm(bbl)

ft2m(ft)

m2ft(m)

hpa2pa(hpa)

pa2hpa(pa)

kpa2pa(kpa)

pa2kpa(pa)

bar2pa(bar)

pa2bar(pa)

atm2bar(atm)
}
\arguments{
\item{Perm_Darcy}{numeric; permeability, in darcies}

\item{Perm_sqm}{numeric; permeability, in \eqn{[m^2]}}

\item{milli}{logical; should the result be divided by 1000?}

\item{bbl}{numeric; US oil barrels (1 bbl ~ 42 gallons US liquid)}

\item{ft}{numeric; length in feet}

\item{m}{numeric; length in meters}

\item{hpa}{numeric; pressure in hecto-Pascals}

\item{pa}{numeric; pressure in Pascals}

\item{kpa}{numeric; pressure in kilo-Pascals}

\item{bar}{numeric; pressure in bars}

\item{atm}{numeric; pressure, in standard atmospheres}
}
\description{
Convert within commonly found units.
}
\details{
There is a panoply of units in hydrogeology literature, so
this represents a very small fraction of possibilities.  There are 
generally three sets of conversion tools here:

\itemize{
\item{length}{ convert meters to feet, etc}
\item{permeability}{ convert from darcies to SI, etc; these will begin with
\code{to_}. For example, \code{to_msquared} converts darcies to \eqn{m^2}.}
\item{pressure}{ convert from pascal to bars, etc}
}
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
\seealso{
\code{\link{dimensional_units}}, \code{\link{hydrogeo.p}}
}

