% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogeo.R
\docType{package}
\name{hydrogeo}
\alias{hydrogeo}
\alias{hydrogeo-package}
\title{Groundwater data presentation and interpretation.}
\description{
Contains one function for drawing Piper (or Piper-Hill) diagrams
from water analyses for major ions.
}
\author{
Myles English \email{myles@rockhead.biz}
}
\keyword{groundwater}
\keyword{hydrogeology}
\keyword{hydrology}
\keyword{package}
\keyword{water}

