% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piperPaper}
\alias{piperPaper}
\title{Create a new piperplot object}
\usage{
piperPaper(size = NULL, ...)
}
\arguments{
\item{size}{integer related to the size of the plot area}

\item{...}{additional arguments, as for \code{\linkS4class{piperplot}}}
}
\description{
Create a new piperplot object
}
\examples{
library(hydrogeo)
p = piperPaper(size=1)
plot(p)
}

