% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\docType{class}
\name{piperplot-class}
\alias{Axis,piperplot-method}
\alias{piperplot-class}
\title{Class \code{piperplot}}
\usage{
\S4method{Axis}{piperplot}(x = NULL)
}
\arguments{
\item{x}{an object of class piperplot}
}
\description{
Objects of this class are plottable as empty (i.e. no points) Piper-Hill
diagrams
}
\section{Methods (by generic)}{
\itemize{
\item \code{Axis}: Add axes to a piperplot
}}
\section{Slots}{

\describe{
\item{\code{size}}{Object of class \code{numeric} --- Length of the (square) plot
area, defaults to 300}

\item{\code{call}}{R call that created it}
}}

