% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\docType{methods}
\name{plot,piperplot-method}
\alias{plot,piperplot-method}
\title{Plot the diagram area with two triangles and a diamond}
\usage{
\S4method{plot}{piperplot}(x, axes = TRUE, ...)
}
\arguments{
\item{x}{object of class piperplot}

\item{axes}{logical saying whether to draw the axes or not, defaults to TRUE}

\item{...}{further arguments to plot.default}
}
\description{
Plot the diagram area with two triangles and a diamond
}
