\docType{data}
\name{shepard}
\alias{shepard}
\title{Shepard's (1954) grain-size classification.}
\description{
  Values are percent grain-size due to sand, silt, and clay
  in sedimentary material.
}
\details{
  For use in ternary plots.
}
\references{
  Shepard, F.P. (1954), Nomenclature based on
  sand-silt-clay ratios, \emph{Journal of Sedimentary
  Petrology}, \strong{24}, p. 151-158.
}
\seealso{
  \code{\link{sand_silt_clay}} and
  \code{\link{shepard_plot}}
}
\keyword{datasets}

