% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_skempton.R
\name{skempton}
\alias{skempton}
\title{Calculate Skempton's coefficient \eqn{B}.}
\usage{
skempton(Beta, Phi, Beta_u = NULL, Beta_f = NULL)
}
\arguments{
\item{Beta}{numeric; compressibility \eqn{\beta}}

\item{Phi}{numeric; \eqn{\Phi}}

\item{Beta_u}{numeric; \eqn{\beta[u]} solid matrix - undrained}

\item{Beta_f}{numeric; \eqn{\beta[f]} fluid}
}
\description{
Skempton related changes in applied stress to changes in fluid pressure; this
is the famous constant \eqn{B}.
}
\details{
Calculates Equation (4) from Rojstaczer and Agnew (1989)
}
\section{Assumptions}{
  

As noted by Rice and Cleary [1976], this definition for 
\eqn{B} assumes that the rock matrix is homogeneous and all the 
pore space is interconnected.
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
\seealso{
\code{\link{compressibility}}, \code{\link{hydrogeo.p}}
}

