% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{testData}
\alias{testData}
\title{Major ions as a percentage of total major ions - Test Data}
\usage{
testData(n)
}
\arguments{
\item{n}{Number of test samples to be generated.}
}
\description{
Major ions as a percentage of total major ions - Test Data
}
\examples{
library(hydrogeo)
lp <- piper( testData(26) )
}
