% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_elasticprops.R
\name{Undrained.Response}
\alias{Undrained.Response}
\title{Calculate undrained pressure response to volume strain}
\usage{
Undrained.Response(mu, B, nu_u = NULL, as.areal = FALSE)
}
\arguments{
\item{mu}{numeric; the shear modulus of the rock, in Pa}

\item{B}{numeric; Skempton's coefficient [0,1]}

\item{nu_u}{numeric; the undrained Poisson's ratio}

\item{as.areal}{logical; should the response be to areal strain
rather than volume strain?  This assumes plane stress (a 
traction free surface) and thus the form of vertical strain.}
}
\description{
Calculate undrained pressure response to volume strain
}

