% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dimensional_units}
\alias{dimensional_units}
\title{Gives dimensional units for quantity desired.}
\usage{
dimensional_units(quantity = c("show.all", "conductivity", "diffusivity",
  "transmissivity", "storativity", "specific.storage", "viscosity", "density",
  "kinematic.viscosity", "permeability", "force", "stress", "strain",
  "compressibility", "porosity", "skempton"), verbose = TRUE)
}
\arguments{
\item{quantity}{character; the quantity to show units for}

\item{verbose}{logical; should a message be printed also?}
}
\value{
A character string with the quantity, and dimensional-units; invisibly.
}
\description{
Poroelasticity and hydrogeology can be a mish-mash of units;
keeping them straight can be challenging.  This returns a
string with the dimensional units, and (optionally) prints them
as a message.
}
\details{
This function uses \code{match.arg} so partial representations are OK.

The dimensional representations this far are
\itemize{
\item{L}{ length}
\item{T}{ time}
\item{M}{ mass}
}
with SI units \eqn{[m]}, \eqn{[s]}, and \eqn{[kg]} respectively.
}
\examples{
### code to be run
print(dimensional_units(verbose=FALSE)) # show all
dimensional_units("diffusivity") # [ L^2 / T ]
dimensional_units("diff") # matches also, because of match.arg
print(dimensional_units("diff", FALSE)) # print, but no message
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
\seealso{
\code{\link{hydrogeo.p-units}}, \code{\link{hydrogeo.p}}
}

