% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_elasticprops.R
\name{Elastic.Properties}
\alias{Elastic.Properties}
\alias{Lame_constant}
\alias{Poissons_ratio}
\alias{Youngs_modulus}
\alias{bulk_modulus}
\alias{elasticProperties}
\alias{elastic_properties}
\alias{shear_modulus}
\title{Elastic properties from seismic velocities}
\usage{
Elastic.Properties(Vp, Vs, dens)

shear_modulus(Vs, dens)

bulk_modulus(Vp, Vs, dens, mu = NULL)

Lame_constant(Vp, Vs, dens)

Poissons_ratio(Vp, Vs)

Youngs_modulus(Vp, Vs, dens)
}
\arguments{
\item{Vp}{numeric; the P-wave velocity of the rock in km/s}

\item{Vs}{numeric; the S-wave velocity of the rock in km/s}

\item{dens}{numeric; the rock density, in kg/m^3}

\item{mu}{numeric; the shear modulus of the rock, in Pa}
}
\description{
Elastic properties from seismic velocities
}

