\docType{data}
\name{hydrogeo-constants}
\alias{hydrogeo-constants}
\title{Constants used in many calculations.}
\format{List of 10
 $ gravity   : num 9.81
 $ fluid_dens: num 1000
 $ dyn_visc  : num 0.001
 $ nu        : num 0.25
 $ nu_u      : num 0.333
 $ VpVs      : num 1.73
 $ Beta_u    : num 2e-11
 $ Beta_f    : num 4.4e-10
 $ atm       :List of 6
 $ sqm2sqcm  : num 10000}
\usage{
  .constants
}
\description{
  Constants used in many calculations.
}
\section{Typical Values}{
  \describe{ \item{Hydraulic conductivity,
  \eqn{C}}{\eqn{1^{-0} -- 1^{-0}}} }
}
\seealso{
  \code{\link{hydrogeo-units}}, \code{\link{hydrogeo}}
}
\keyword{datasets}

