\name{hydrogeo-constants}
\alias{hydrogeo-constants}
\title{Constants used in many calculations.}
\description{
  Constants used in many calculations.
}
\section{Typical Values}{
  \describe{ \item{Hydraulic conductivity,
  \eqn{C}}{\eqn{1^{-0} -- 1^{-0}}} }
}
\seealso{
  \code{\link{hydrogeo-units}}, \code{\link{hydrogeo}}
}

