% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper}
\alias{piper}
\title{Create a new piper object}
\usage{
piper(d, ...)
}
\arguments{
\item{d}{list passed to class piper, \code{\linkS4class{piper}}}

\item{...}{additional arguments, as for \code{\linkS4class{piper}}}
}
\description{
Create a new piper object
}

